+++ cmd/frontend/internal/db/mockstores.go
@@ -19,4 +19,6 @@ type MockStores struct {
 	Phabricator MockPhabricator
 
 	ExternalAccounts MockExternalAccounts
+
+	OrgInvitations MockOrgInvitations
 }