+++ cmd/frontend/internal/db/users.go
@@ -354,6 +354,9 @@ func (u *users) Delete(ctx context.Context, id int32) error {
 	if _, err := tx.ExecContext(ctx, "UPDATE user_external_accounts SET deleted_at=now() WHERE user_id=$1 AND deleted_at IS NULL", id); err != nil {
 		return err
 	}
+	if _, err := tx.ExecContext(ctx, "UPDATE org_invitations SET deleted_at=now() WHERE deleted_at IS NULL AND (sender_user_id=$1 OR recipient_user_id=$1)", id); err != nil {
+		return err
+	}
 
 	return nil
 }