+++ schema/site_stringdata.go
@@ -386,11 +386,6 @@ const SiteSchemaJSON = `{
       "description": "Java: The password to authenticate to the private Artifactory.",
       "type": "string"
     },
-    "secretKey": {
-      "description":
-        "A base64-encoded secret key for this site, used for generating links to invite users to organizations. If you have ` + "`" + `openssl` + "`" + ` installed, you can generate a valid key by running ` + "`" + `openssl rand -base64 32` + "`" + `.",
-      "type": "string"
-    },
     "auth.provider": {
       "description":
         "The authentication provider to use for identifying and signing in users. Defaults to \"builtin\" authentication.\n\nDEPRECATED: Use \"auth.providers\" instead. During the deprecation period (before this property is removed), provider set here will be added as an entry in \"auth.providers\".",