+++ web/src/org/OrgAvatar.tsx
@@ -7,5 +7,9 @@ export const OrgAvatar: React.SFC<{
     /** The organization's name. */
     org: string
 
+    size?: 'md' | 'lg'
+
     className?: string
-}> = ({ org, className = '' }) => <div className={`org-avatar ${className}`}>{org.substr(0, 2).toUpperCase()}</div>
+}> = ({ org, size = 'md', className = '' }) => (
+    <div className={`org-avatar org-avatar--${size} ${className}`}>{org.substr(0, 2).toUpperCase()}</div>
+)