+++ web/src/user/settings/RedirectToUserSettings.tsx
@@ -15,13 +15,11 @@ export const RedirectToUserSettings: React.SFC<{
     // If not logged in, redirect to sign in
     if (!user) {
         const newURL = new URL(window.location.href)
-        newURL.pathname = location.pathname === '/settings/accept-invite' ? '/sign-up' : '/sign-in'
+        newURL.pathname = '/sign-in'
         // Return to the current page after sign up/in.
         newURL.searchParams.set('returnTo', window.location.href)
         return <Redirect to={{ pathname: newURL.pathname, search: newURL.search }} />
     }
 
-    // Append full location.pathname instead of just "/settings" so that we support /settings/accept-invite
-    // redirect, too.
-    return <Redirect to={{ pathname: `${userURL(user.username)}${location.pathname}`, search: location.search }} />
+    return <Redirect to={{ pathname: `${userURL(user.username)}/settings`, search: location.search }} />
 }